tests/requirements.txt
# Core dependencies
langgraph>=0.0.40
langchain>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Async & HTTP
aiohttp>=3.9.0
tenacity>=8.2.0

# Email parsing
extract-msg>=0.45.0

# Document processing - REAL IMPLEMENTATIONS
PyMuPDF>=1.23.0  # For PDF parsing
python-docx>=1.0.0  # For DOCX (optional, Azure can handle)
Pillow>=10.0.0  # For image processing

# Azure Document Intelligence
azure-ai-formrecognizer>=3.3.0
azure-core>=1.29.0
azure-identity>=1.15.0

# LLM providers
openai>=1.10.0  # For OpenAI GPT-4

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging & Monitoring
structlog>=23.1.0

